+++ src/Google/Translate.hs
@@ -2,11 +2,17 @@
 
 module Google.Translate (translate) where
 
-import Control.Lens ((&), (^.), (^?), (.~), toListOf)
+import Control.Lens
 import Data.Aeson
 import GHC.Generics (Generic)
-import Network.Wreq
-import Data.Map (Map)
+import Network.Wreq as W
+--import Network.Wreq.Lens
+--import Data.Map (Map)
+import Control.Exception as E
+import Network.HTTP.Client as C
+
+import qualified Data.ByteString.Char8 as BSC
+import qualified Data.ByteString.Lazy  as LBS
 
 -- Data sent to google 
 
@@ -41,9 +47,21 @@ instance FromJSON BodyData where
   parseJSON (Object v) =
     BodyData <$> v .: "data"
 
-translate :: String -> String -> String -> String -> IO String
+translate :: String -> String -> String -> String -> IO (Either String String)
 translate key src dest text = do
-    r <- post ("https://translation.googleapis.com/language/translate/v2?key="++key) (toJSON $ PostData text src dest "text")
-    case (decode (r ^. responseBody) :: Maybe BodyData) of 
-        Nothing -> return "invalid response"
-        Just v -> return $ translatedText . head . translations . data' $ v
+    eR <- internalPost (PostData text src dest "text") key
+    case eR of 
+        Left s -> return $ Left s
+        Right r ->
+            -- TODO get a better error if decode failed
+            case (decode (r ^. W.responseBody) :: Maybe BodyData) of 
+                Nothing -> return $ Left "can't decode the body..."
+                Just v -> return $ Right $ translatedText . head . translations . data' $ v
+    where
+        internalPost :: PostData -> String -> IO (Either String (Response LBS.ByteString))
+        internalPost pd key = do 
+            (Right <$> post ("https://translation.googleapis.com/language/translate/v2?key="++key) (toJSON pd)) `E.catch` handler
+        handler :: HttpException -> IO (Either String (Response LBS.ByteString))
+        handler e@(HttpExceptionRequest _ (StatusCodeException r' _)) =
+            return $ Left $ BSC.unpack $ r' ^. W.responseStatus . W.statusMessage 
+