+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Revision history for tp-eo
 
-## 0.0.20.0 -- 2020-04-26
+## 0.0.21.0 -- 2020-05-04
+* enhance error management when calling Google.Translate.translate
+* add a Servant API and server
 
+## 0.0.20.1 -- 2020-04-30
+* A first pre-release
+* minor fix and enhancemment 
+
+## 0.0.20.0 -- 2020-04-26
 * First version of the command line tp-eo
 * Cabal -> Stack
 
 ## 0.0.19.0 -- 2020-04-25
-
 * First version of the module
 * Supports the 19 lessons of jan Pije (test)